+++ applications/diem-core/.npmrc
@@ -1,0 +1,1 @@
+legacy-peer-deps=true
