+++ applications/diem-core/src/webpack/src/webpack.front.ts
@@ -13,9 +13,9 @@ import { AngularWebpackPlugin } from '@ngtools/webpack';
 import { InjectManifest } from 'workbox-webpack-plugin';
 /** const Ba = require('webpack-bundle-analyzer'); */
 
-console.info(`$webpack.front: environment: ${process.env.webpackenv}`);
+import { splitChunks, patterns, assets } from './webpack.common';
 
-const assets: string = 'public/assets/[name].[hash].[ext]';
+console.info(`$webpack.front: environment: ${process.env.webpackenv}`);
 
 interface IntEnv {
     APPCOOKIE?: string;
@@ -48,13 +48,14 @@ module.exports = {
 
     output: {
         chunkFilename: 'public/js/[name][chunkhash].js',
-        filename: 'public/js/bundle[chunkhash].js',
+        filename: 'public/js/bundle.[chunkhash].js',
         path: `${(global as any).__basedir}/`,
         publicPath: `${env.APPPATH}/`,
         assetModuleFilename: assets,
     },
 
     optimization: {
+        runtimeChunk: 'single',
         moduleIds: 'deterministic',
         minimize: true,
         minimizer: [
@@ -70,27 +71,7 @@ module.exports = {
             }),
             new CssMinimizerPlugin(),
         ],
-        splitChunks: {
-            chunks: 'async',
-            minSize: 20000,
-            minRemainingSize: 0,
-            minChunks: 1,
-            maxAsyncRequests: 30,
-            maxInitialRequests: 30,
-            enforceSizeThreshold: 50000,
-            cacheGroups: {
-                defaultVendors: {
-                    test: /[\\/]node_modules[\\/]/,
-                    priority: -10,
-                    reuseExistingChunk: true,
-                },
-                default: {
-                    minChunks: 2,
-                    priority: -20,
-                    reuseExistingChunk: true,
-                },
-            },
-        },
+        splitChunks,
     },
 
     module: {
@@ -175,27 +156,7 @@ module.exports = {
 
     plugins: [
         new CopyWebpackPlugin({
-            patterns: [
-                {
-                    from: 'src/server/index.pug',
-                    to: 'server/index.pug',
-                },
-                {
-                    from: `${(global as any).__basedir}/public/images/*.jpg`,
-                },
-                {
-                    from: `${(global as any).__basedir}/public/images/*.gif`,
-                },
-                {
-                    from: `${(global as any).__basedir}/public/images/favicon*.*`,
-                },
-                {
-                    from: `${(global as any).__basedir}/public/images/diem_logo.*`,
-                },
-                {
-                    from: `${(global as any).__basedir}/public/images/diem_s.*`,
-                },
-            ],
+            patterns,
         }),
 
         new AngularWebpackPlugin({
